requirements.txt
numpy==1.21.6
scipy==1.11.2
matplotlib==3.7.2
Pillow==9.4.0
opencv-python==4.8.0.76
tqdm==4.64.1


accelerate==0.22.0
transformers==4.33.1
diffusers==0.20.2
huggingface-hub==0.16.4


open3d==0.17.0
trimesh==3.23.5
kornia==0.6.12
einops==0.6.1


gradio==3.43.2
fastapi==0.103.1
uvicorn==0.23.2

omegaconf==2.3.0
PyMCubes==0.1.6
plyfile==1.1


./extension/speed_up
./extension/mask_control2
./extension/diff-surfel-rasterization
./extension/max_cos
./extension/simple-knn